+++ CHANGELOG.rst
@@ -2,6 +2,23 @@
 Changelog
 #########
 
+----------------------
+2021.10.0 - 2021-10-01
+----------------------
+
+    * packaging
+        * change version numbering scheme to ``YYYY.MM.x``
+
+    * core
+        * add environment configuration via MHUB_HOST, MHUB_USER, MHUB_PASSWORD environment variables.
+
+    * CLI
+        * smooth progress bar
+        * better error handling
+
+    * testing
+        * use newest mhub release for testing
+
 ----------------
 0.2 - 2021-09-23
 ----------------